Implement custom newline normalization.
Instead of reading file in text mode, read it in binary and convert CRLF to LF,
and any stray CRs to LF (XML EOL handling rules). This ensures that the generated
data is cross-platform and well-behaved for foreign sources - i.e. search results
are correct on Windows for source data with Mac line endings.