Improve build output by printing starting statistics first.
This eliminates leftover "building file table" message for empty projects and
removes this message faster when the first file to process is huge. The result
is that the user does not have the wrong impression that file table building
takes too long even in case of empty projects/huge files.