Terminate search as soon as output limit is reached
Previously we processed a chunk fully until we hit the output limit.
Moreover, multiple chunks could be read and queued for decompression, even if
the limit was reached very quickly.

This was especially problematic on worst-case matches where searching takes a lot of time -
for example for "." queries with highlighting we need to scan an entire line for ".".

Terminating earlier reduces search time with highlighting for "." with L100 from 9500 ms to 130 ms.