Command-T: Adjust score to favor word boundary matches
We assign special weights to first and last character in the match, with
score that depends on the neighboring character - string boundary is best,
followed by a path element boundary, followed by a word boundary.

This allows us to prefer e.g. common.h to BulletCommonDynamics.h when the
filter is "common".