Command-T: Fix rightScore for last character
Since we were reducing the range the last character would always fail the
range check. Rename offset to begin and size to end to fix this.

Also skip character that don't occur in the pattern early; this used to hurt
performance since the code was branchless but now it's worth skipping the extra
score computations.