Prevent possible (but unlikely) buffer overrun
The loop in literal matcher does not implicitly guarantee that the final
memcmp is in-bounds.

It's cheap to perform the check before doing memcmp since it needs the same
set of variables and memcmp almost always succeeds anyway. Adjusting the loop
bounds may make the loop condition less efficient and we care more about that.