Switch to Unicode file operations on Windows
Instead of using ANSI, use UTF16 APIs for file access and convert paths
assuming the "cross-platform" format for path storage is UTF-8 (which
does not necessarily handle all corner cases like paths with invalid
surrogates but should be good enough).

This also means that file paths for matches are output using UTF-8 which
needs a UTF8 terminal to display correctly.