Keep directories open during file traversal on POSIX
This results in significant speedup (9.7s -> 3.9s) of traversal on Windows
Subsystem for Linux for the traversal of Linux kernel source tree. The bulk
of the traversal time is spent doing stat calls so it could be some sort of fd
cache and fd being kept open accelerating the process.