Optimize chunk update by doing partial data decompression
Chunks that contain no file changes can be generally reused without
recompression; this also means we don't actually need the file data, but
we still need the file table. Since decompression of these chunks is a
significant cost during the update, add a field to the header that
indicates the file table size and decompress the file table separately
from chunk contents.

This improves the no-op incremental build for Linux kernel tree pack (on
a hot cache) from 1.1s to 0.6s. There is a small cost for decompressing
the file table redundantly for chunks we do need to process but it's
dwarfed by file I/O and chunk recompression so it's not worth getting
rid of.