Improve openFile function to handle long paths and special names
WinAPI is pretty bizarre. To opt into long paths and to be able to open
files like aux.c (even using a full path), you have to prefix your path
with \\?\. Unfortunately, relative paths do not work in this format, so
we have to construct a full path for this to work. Since path
normalization is not free, we only do this for relative paths.

Additionally optimize normalizePath by eliminating string allocations.