build: Add last file path to the chunk as extra data
To implement changed file lists and maintain search ordering, we'll need
to be able to search in the list of all files for the necessary range.
This requires knowledge about neighboring chunks, which is cumbersome to
maintain during incremental builds, but it's trivial to establish this
at search time if the path is readily available, since search dispatches
chunks one at a time.

Currently this information is only available in the compressed section
at an undetermined offset, and you'd have to decompress the chunk file
table to get this which is a non-starter since it'd introduce
significant serial cost for searches. Instead we'll just store this
information redundantly in an uncompressed form.

In the future we can add other information to the extra block if
necessary, assuming it's small, has a variable size and does not require
much processing.