watch: Implement change monitoring
Whenever a file changes we now update the .qgc file with a new list of
changed files as long as the file matches the project group filters.

There is currently no throttling on the .qgc update so this may be
suboptimal for large batches of operations such as git checkout.
Additionally, the starting point is currently an empty set of files so
repeatedly running watch will not work well. This will be addressed
later.