watch: Implement initial change list retrieval
When watch starts we now read the data pack, extract the file table and
compare it with the file table we got via filesystem scan to determine
the initial set of changes to write to .qgc file.

This is basically the same process that update performs, but implemented
somewhat differently.

This code is in dire need of refactoring and it's interesting to
consider reusing update code for this, but it seems to work.

File removals are not currently handled because it's not clear if it's a
good idea to write them to changed list.