search: Optimize processing large change lists
changes vector was being copied for every chunk, which is very expensive
because it causes many allocations. Also add std::move to WorkQueue
implementation to reduce the cost of copying function objects (which is
generally not noticeable but this helps just in case we do need to copy).

On Linux kernel tree with *all* files in .qgc list (46k paths), before
this change the search time is ~11.4 seconds; after std::move it drops
down to 5.2 seconds; after &changes it drops down to 0.4 seconds which
is much more reasonable (the pack search is 0.27 seconds)