search: Fix first-file-is-suffix handling wrt changed files
First file in a chunk may be a suffix of the file in the previous chunk
(or some chunk before that for really big files). We split the changed
files into disjoint sequences so the suffixes weren't correctly ignored
and instead were processed from the pack file even if the file has
changed.

Since we want to only process each changed file once we have to do it as
part of the original chunk. Fortunately we can exploit the knowledge
that the changed file list is ordered so if our file is a suffix our
current change list position is the next entry after our file in the
change list if it exists there at all.

As such we just check if the previous entry in the change list matches
our path for suffixes and skip processing (assuming that the first chunk
where the file occurred with startLine=0 handled the file).