watch: Make it possible to watch multiple projects
Since each watch process never exits at this time we need to watch each
project from a separate thread.

The output from multiple threads gets intermixed in interesting ways
especially because we use \r progress status updates, but it's not that
big of a problem as long as you work on one project at a time.