watch: Amortize change writes in a 1-second interval
Instead of immediately writing accumulated changes when we get a change
request, wait for up to 1 second to gather more changes and then write
all of them.

This bounds the frequency at which we write out change files to reduce
the CPU and I/O cost of doing so. The flow is different from the similar
update functionality - in update, if the changes keep coming we want to
keep delaing the update call - possibly indefinitely! - whereas with
write, if the changes keep coming we still write them out.