watch: Implement update error handling
When the update fails, we need to restore our previous state and try
again. This means restoring the list of changed files to include all
previously changed files (as well as all new changes if any!) and
writing out the change list again because update removes it in the
beginning.

After that we just wait for another stable point in time to be reached.