watch: Always write changes file
Removing the file has an issue - if we run watch and the file list is
empty, we can try to remove a non-existent file which will generate an
error.

Update removes change file anyway, and searching with an empty change
list is not really slower than searching without one - the only extra
overhead is opening an empty file - so for now just change the code to
always write the file even if it's empty.