change: Implement new change command
Given a list of file names/paths, it normalizes them to get full paths
and then proceeds to add them to the change list of a project if the
path matches one of the project groups.

It's possible to request a change to all projects using usual */%
project masks, which would be a good generic way to handle editor file
changes.