Fix assertion in incremental update for empty files
In a few places in the code, we used pendingSize > 0 to gate chunk
updates; if any input files are empty, this check is incorrect since it
ignores the pending files. This can lead to assertions as well as files
missing from the chunk index.

This change refactors the chunk build logic to fix this:

- Remove all flush* helpers and instead explicitly inline the logic into
the appropriate functions

- When adding a chunk, use a simpler loop that writes chunks of
appropriate sizes up until the pending set is empty or the chunk size is
too small

- When finishing the build, make sure to write all files including the
empty ones

This change also adds comments to existing size-based logic to make it
more clear why it's correct.

Fixes #12.