Update heuristics to match old logic
The new code results in suboptimal behavior because it will split 1.6x
into 1x and 0.6x which results in a chunk that's too small which will
need to be recompressed.

Instead we now use the code to determine the chunk size using the old
rules (chunks between 1.5x and 2x are split in half) to reduce chunk
recompression.