Fix an off-by-one error in literal matcher for matches at the end
When a match was supposed to last until the very end of the string, an
off-by-one error in LiteralMatcher16 prevented us from confirming the
match inside the SIMD loop.

The correct condition here is <= since we want to know if it's safe to
read the pattern up until the end of the string, not if it's safe to
index past that.