update: Streamline special handling of first file in a chunk
The previous logic was correct but not obviously so: when processing the first
file in a chunk, we had to compare it to the last written file in the list. We
would subsequently process the first file again in a loop; the logic worked out
because of an extra if() that skipped files with startLine != 0, but this way
it's more obvious that the code is correct.