Update file data compression level to 3 (from 5)
This results in ~3% larger files but the build throughput is ~35% higher.

The original compression level has been 9; it was changed from 9 to 5 in
2017, for 1% regression in size vs 50% gain in throughput. This change
represents a more modest gain in throughput and a larger regression in size,
but the query performance is generally not impacted here either.

This should provide a little bit of a better balance on modern systems where
it takes a lot of cores for compression to beat the disk read throughput.