extern: Update re2 to 2023-03-01
This might be the last update of re2 dependency for this project, as more recent
versions introduce a dependency on abseil which is against the general policy
for qgrep.

The good news is that this update seems to significantly increase regex matching
performance in some cases; it probably won't very notably affect many queries as
we have many optimizations that try to ensure we don't go to regex matching, but
some queries, when forced to match, seem to run ~3x faster on macOS/M2.