Further cleanup of macOS build setup
I don't think we need force_cpusubtype_ALL at this point or stdlib; CMake
build runs into the same problem with stdlib flag as make so it's easier
to just remove it.

We do need macOS version setting, both to ensure a reasonable binary target,
and because we use an API that's deprecated in macOS 13.